Move help method out of conftest and add tests for subqueries
