+++ website/python/test_brewmonitor/admin/test_project_views.py
@@ -3,7 +3,7 @@ from http import HTTPStatus
 import pytest
 from brewmonitor.storage.tables import Project
 from flask import url_for
-from test_brewmonitor.conftest import config_from_client
+from test_brewmonitor.utils import config_from_client
 
 
 pages = (
