+++ website/python/test_brewmonitor/admin/test_sensor_views.py
@@ -4,7 +4,7 @@ import attr
 import pytest
 from brewmonitor.storage.tables import Sensor
 from flask import url_for
-from test_brewmonitor.conftest import config_from_client
+from test_brewmonitor.utils import config_from_client
 
 
 pages = (
