+++ website/python/test_brewmonitor/admin/test_user_views.py
@@ -3,7 +3,7 @@ from http import HTTPStatus
 import pytest
 from brewmonitor.storage.tables import User
 from flask import url_for
-from test_brewmonitor.conftest import config_from_client, find_user
+from test_brewmonitor.utils import config_from_client, find_user
 
 
 pages = (
