+++ website/python/test_brewmonitor/conftest.py
@@ -1,41 +1,12 @@
-import os
-from copy import deepcopy
-from datetime import datetime
 from http import HTTPStatus
 from tempfile import NamedTemporaryFile
-from typing import Optional
 
 import pytest
-import yaml
-from brewmonitor.app import make_app
-from brewmonitor.configuration import Configuration
-from brewmonitor.storage.tables import Project, Sensor, User
+from brewmonitor.storage.tables import Project, Sensor
 from flask import Flask
 from make_dummy_data import make_dummy_data
-
-
-test_config = {
-    'flask configuration': {
-        'TESTING': True,
-        'SERVER_NAME': 'unit-test-brewmonitor.domain',
-    },
-}
-
-
-def config_from_client(client_obj: Flask) -> Configuration:
-    return client_obj.config['brewmonitor config']
-
-
-def make_clean_client(config_file: NamedTemporaryFile, db_file: NamedTemporaryFile) -> Flask:
-    apply_config = deepcopy(test_config)
-    apply_config['sqlite file'] = db_file.name
-
-    yaml.dump(apply_config, config_file.file, encoding='utf-8')
-
-    os.environ['BREWMONITOR_CONFIG'] = config_file.name
-
-    print(f'Making app with config={config_file.name} and db={db_file.name}')
-    return make_app('unit test secrets')
+from test_brewmonitor.constants import preset_when
+from test_brewmonitor.utils import config_from_client, find_project, find_sensor, find_user, make_clean_client
 
 
 @pytest.fixture(scope='function')
@@ -44,9 +15,6 @@ def tmp_app() -> Flask:
     with NamedTemporaryFile() as config_file:
         with NamedTemporaryFile() as db_file:
             yield make_clean_client(config_file, db_file)
-
-
-preset_when = datetime(2021, 11, 26, 11, 34)
 
 
 @pytest.fixture(scope='package')
@@ -70,53 +38,6 @@ def public_client(preset_app):
         with preset_app.test_client() as client:
             print('Public client')
             yield client
-
-
-def find_user(db_conn, username: str) -> Optional[User]:
-    cursor = db_conn.execute(
-        """
-        select id
-        from User
-        where username=?
-        """,
-        (username,),
-    )
-    # cursor.row_factory = User.row_factory
-    user = cursor.fetchone()
-    if user:
-        return User.find(db_conn, user[0])
-    return None
-
-
-def find_sensor(db_conn, sensor_name: str) -> Optional[Sensor]:
-    cursor = db_conn.execute(
-        """
-        select id
-        from Sensor
-        where name=?
-        """,
-        (sensor_name,),
-    )
-    # cursor.row_factory = Sensor.row_factory
-    sensor = cursor.fetchone()
-    if sensor is not None:
-        return Sensor.find(db_conn, sensor[0])
-    return None
-
-
-def find_project(db_conn, project_name: str) -> Optional[Project]:
-    cursor = db_conn.execute(
-        """
-        select id
-        from Project
-        where name=?
-        """,
-        (project_name,),
-    )
-    project = cursor.fetchone()
-    if project is not None:
-        return Project.find(db_conn, project[0])
-    return None
 
 
 @pytest.fixture
