+++ website/python/test_brewmonitor/constants.py
@@ -1,0 +1,10 @@
+from datetime import datetime
+
+
+preset_when = datetime(2021, 11, 26, 11, 34)
+test_config = {
+    'flask configuration': {
+        'TESTING': True,
+        'SERVER_NAME': 'unit-test-brewmonitor.domain',
+    },
+}
